Change past events to status: past
There needs to be a better mechanism for catching events by
passed dates and changing them appropriately.

It could be done with javascript, but adds a js dependency to
view the page correctly.

It could be coupled with a build-time method, but unless the site
is constantly built on a schedule (as opposed to a push on git)
this won't really fix it either.